flask-server/pyproject.toml
[tool.poetry]
name = "flask-server"
version = "0.1.0"
description = ""
authors = ["hvieira"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.13"
Flask = "^3.0.3"
Flask-SQLAlchemy = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
coverage = "^7.6.1"